requirements.txt
# Основные зависимости
--extra-index-url https://download.pytorch.org/whl/cu124
torch==2.6.0
torchvision==0.21.0
torchaudio==2.6.0
transformers==4.51.3
datasets==2.11.0
accelerate==1.6.0
sklearn-compat==0.1.3
sklearn-pandas==2.2.0